Rearrange the headers so that telescope.hpp can be used in other projects to provide the functions.
* src/read_kallisto_alignments.cpp:
