--- include/read_themisto_alignments.hpp
@@ -1,19 +1,0 @@
-#ifndef TELESCOPE_READ_THEMISTO_ALIGNMENTS_HPP
-#define TELESCOPE_READ_THEMISTO_ALIGNMENTS_HPP
-
-#include <cstddef>
-#include <vector>
-#include <fstream>
-
-#include "common.hpp"
-
-namespace telescope {
-namespace read {
-void Themisto(const Mode &mode, const uint32_t n_refs, std::vector<std::istream*> &streams, CompressedAlignment *aln);
-void ThemistoGrouped(const Mode &mode, const std::vector<uint16_t> &group_indicators, const uint32_t n_refs, const uint16_t n_groups, std::vector<std::istream*> &streams, GroupedAlignment *aln);
-void ThemistoAlignedReads(const Mode &mode, const uint32_t n_refs, std::vector<std::istream*> &streams, ThemistoAlignment *aln);
-void ThemistoToKallisto(const Mode &mode, const uint32_t n_refs, std::vector<std::istream*> &streams, KallistoAlignment *aln);
-}
-}
-
-#endif
